Minor code optimizations

+++ public_html/index.php
@@ -159,30 +159,32 @@ if ( $_VARS['totalhits'] % 100 === 0 ) {
     COM_rdfUpToDateCheck();
 }
 
-// Scan for any stories that have expired and should be archived or deleted
-$asql = "SELECT sid,tid,title,expire,statuscode FROM {$_TABLES['stories']} ";
-$asql .= 'WHERE (expire <= NOW()) AND (statuscode = ' . STORY_DELETE_ON_EXPIRE;
-if (empty ($archivetid)) {
-    $asql .= ')';
-} else {
-    $asql .= ' OR statuscode = ' . STORY_ARCHIVE_ON_EXPIRE . ") AND tid != '".DB_escapeString($archivetid)."'";
-}
-$expiresql = DB_query ($asql);
-while (list ($sid, $expiretopic, $title, $expire, $statuscode) = DB_fetchArray ($expiresql)) {
-    if ($statuscode == STORY_ARCHIVE_ON_EXPIRE) {
-        if (!empty ($archivetid) ) {
-            COM_errorLOG("Archive Story: $sid, Topic: $archivetid, Title: $title, Expired: $expire");
-            DB_query ("UPDATE {$_TABLES['stories']} SET tid = '".DB_escapeString($archivetid)."', frontpage = '0', featured = '0' WHERE sid='".DB_escapeString($sid)."'");
+if ( $_VARS['totalhits'] % 50 === 0 ) {
+    // Scan for any stories that have expired and should be archived or deleted
+    $asql = "SELECT sid,tid,title,expire,statuscode FROM {$_TABLES['stories']} ";
+    $asql .= 'WHERE (expire <= NOW()) AND (statuscode = ' . STORY_DELETE_ON_EXPIRE;
+    if (empty ($archivetid)) {
+        $asql .= ')';
+    } else {
+        $asql .= ' OR statuscode = ' . STORY_ARCHIVE_ON_EXPIRE . ") AND tid != '".DB_escapeString($archivetid)."'";
+    }
+    $expiresql = DB_query ($asql);
+    while (list ($sid, $expiretopic, $title, $expire, $statuscode) = DB_fetchArray ($expiresql)) {
+        if ($statuscode == STORY_ARCHIVE_ON_EXPIRE) {
+            if (!empty ($archivetid) ) {
+                COM_errorLOG("Archive Story: $sid, Topic: $archivetid, Title: $title, Expired: $expire");
+                DB_query ("UPDATE {$_TABLES['stories']} SET tid = '".DB_escapeString($archivetid)."', frontpage = '0', featured = '0' WHERE sid='".DB_escapeString($sid)."'");
+                CACHE_remove_instance('story_'.$sid);
+                CACHE_remove_instance('whatsnew');
+            }
+        } else if ($statuscode == STORY_DELETE_ON_EXPIRE) {
+            COM_errorLOG("Delete Story and comments: $sid, Topic: $expiretopic, Title: $title, Expired: $expire");
+            STORY_deleteImages ($sid);
+            DB_query("DELETE FROM {$_TABLES['comments']} WHERE sid='".DB_escapeString($sid)."' AND type = 'article'");
+            DB_query("DELETE FROM {$_TABLES['stories']} WHERE sid='".DB_escapeString($sid)."'");
             CACHE_remove_instance('story_'.$sid);
             CACHE_remove_instance('whatsnew');
         }
-    } else if ($statuscode == STORY_DELETE_ON_EXPIRE) {
-        COM_errorLOG("Delete Story and comments: $sid, Topic: $expiretopic, Title: $title, Expired: $expire");
-        STORY_deleteImages ($sid);
-        DB_query("DELETE FROM {$_TABLES['comments']} WHERE sid='".DB_escapeString($sid)."' AND type = 'article'");
-        DB_query("DELETE FROM {$_TABLES['stories']} WHERE sid='".DB_escapeString($sid)."'");
-        CACHE_remove_instance('story_'.$sid);
-        CACHE_remove_instance('whatsnew');
     }
 }
